# contracts/isyara-ai-services-openapi.yaml
openapi: 3.1.0
info:
  title: Isyara AI Services API
  version: 0.2.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    HTTP contracts for Isyara AI services. STT, TTS, Gloss, and Recall are
    internal Application API dependencies. Sign prediction is an edge-local
    exception called by the web client over loopback on the signer's laptop;
    sign video never reaches the Application API.
servers:
  - url: /
    description: Current AI service origin
security:
  - InternalApiKey: []
...
paths:
...
  /v1/sign/predict:
    post:
      tags: [Sign]
      operationId: predictSign
      summary: Predict one isolated sign from one short video clip.
      description: |
        Called by the signer web client against the loopback Sign Service. The
        browser-local demo mode does not require an internal API key.
      security: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [video, topK, vocabularyVersion]
              properties:
                video:
                  type: string
                  format: binary
                topK:
                  type: integer
                  minimum: 1
                  maximum: 5
                  default: 3
                vocabularyVersion:
                  type: string
                  minLength: 1
                  maxLength: 80
      responses:
        "200":
          description: Prediction completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignPredictionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
  /v1/gloss/normalize:
    post:
      tags: [Gloss]
      operationId: normalizeGloss
      summary: Normalize confirmed sign tokens into final text.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NormalizeGlossRequest"
      responses:
        "200":
          description: Tokens normalized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NormalizeGlossResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
components:
  securitySchemes:
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: true
      description: Caller-generated UUID used for correlation.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    SignCandidate:
      type: object
      additionalProperties: false
      required: [label, confidence]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 80
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
    SignPredictionStatus:
      type: string
      enum: [CONFIDENT, AMBIGUOUS, UNKNOWN, INVALID_INPUT, NO_SIGN]
    SignPredictionResponse:
      type: object
      additionalProperties: false
      required:
        - requestId
        - predictionId
        - prediction
        - confidence
        - candidates
        - status
        - requiresConfirmation
        - modelVersion
        - vocabularyVersion
        - latencyMs
      properties:
        requestId:
          type: string
          format: uuid
        predictionId:
          type: string
          minLength: 1
        prediction:
          type: [string, "null"]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        candidates:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/SignCandidate"
        status:
          $ref: "#/components/schemas/SignPredictionStatus"
        requiresConfirmation:
          type: boolean
        modelVersion:
          type: string
        vocabularyVersion:
          type: string
        latencyMs:
          type: integer
          minimum: 0
...
    ConfirmedSignToken:
      type: object
      additionalProperties: false
      required: [id, label, confirmedAt]
      properties:
        id:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
          maxLength: 80
        confirmedAt:
          type: string
          format: date-time
    NormalizeGlossRequest:
      type: object
      additionalProperties: false
      required: [utteranceId, language, tokens]
      properties:
        utteranceId:
          type: string
          minLength: 1
        language:
          type: string
          default: en
        tokens:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/ConfirmedSignToken"
    NormalizeGlossResponse:
      type: object
      additionalProperties: false
      required: [requestId, utteranceId, text, method, sourceTokenIds, warnings, latencyMs]
      properties:
        requestId:
          type: string
          format: uuid
        utteranceId:
          type: string
        text:
          type: string
        method:
          type: string
          enum: [template, llm, identity]
        sourceTokenIds:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        latencyMs:
          type: integer
          minimum: 0
...
    ErrorBody:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable, requestId, details]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            requestId:
              type: string
              format: uuid
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    PayloadTooLarge:
      description: Request payload exceeds configured limits.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    UnsupportedMediaType:
      description: Audio media type or encoding is unsupported.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    UnprocessableEntity:
      description: Request is valid JSON but cannot be processed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    RateLimited:
      description: Service or upstream provider rate limit reached.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
    ServiceUnavailable:
      description: Model or upstream provider is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    GatewayTimeout:
      description: Upstream inference exceeded the deadline.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"